[ADD] blur view 추가 (#23)
