+++ Maddori.Apple/Maddori.Apple/Screen/StartReflection/UIComponent/CustomVisualEffectView.swift
@@ -1,0 +1,41 @@
+//
+//  CustomVisualEffectView.swift
+//  Maddori.Apple
+//
+//  Created by Mingwan Choi on 2022/10/23.
+//
+
+import UIKit
+
+final class CustomVisualEffectView: UIVisualEffectView {
+    
+    // MARK: - property
+    
+    private let theEffect: UIVisualEffect
+    private let customIntensity: CGFloat
+    private var animator: UIViewPropertyAnimator?
+    
+    // MARK: - life cycle
+    
+    init(effect: UIVisualEffect, intensity: CGFloat) {
+        theEffect = effect
+        customIntensity = intensity
+        super.init(effect: nil)
+    }
+    
+    required init?(coder aDecoder: NSCoder) { nil }
+    
+    deinit {
+        animator?.stopAnimation(true)
+    }
+    
+    override func draw(_ rect: CGRect) {
+        super.draw(rect)
+        effect = nil
+        animator?.stopAnimation(true)
+        animator = UIViewPropertyAnimator(duration: 1, curve: .linear) { [unowned self] in
+            self.effect = theEffect
+        }
+        animator?.fractionComplete = customIntensity
+    }
+}
